added workaround for problem pasting multiple commands
This isn't great. The issue was that the interpreter spawned one thread for each pasted command even from a large list, despite those commands not really needing to be excecuted in parallel. For now the workaround simply sets the started program as the foreground one, preventing others from starting. This does however swallow all other commands from the potentially pasted list, this needs to be implemented better